InstallAndTest/requirements.txt
tensorflow-gpu
keras
h5py
matplotlib
numpy
jupyter
